requirements.txt
hdbscan==0.8.28
matplotlib==3.6.0
numpy==1.23.5
pandas==1.5.0
PySoundFile==0.9.0.post1
scikit_learn==1.1.2
scipy==1.9.1
sounddevice==0.4.6
soundfile==0.11.0
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.64.1
umap_learn==0.5.3
